move UpdateReactionTypeViewModel to commons
